services/auth-service/requirements.txt
# ── Auth Service Dependencies ──────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# ── Config & Validation ───────────────────────────────────────
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# ── Auth & Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ── Cache ─────────────────────────────────────────────────────
redis==5.1.1

# ── Messaging ─────────────────────────────────────────────────
aiokafka==0.11.0

# ── HTTP Client (servisler arası iletişim) ────────────────────
httpx==0.27.2
